requirements.txt
numpy==1.26.4
pillow==10.4.0
pytorch-fid==0.3.0
pytorch-msssim==1.0.0
PyWavelets==1.8.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
torch==2.4.1
torchvision==0.19.1
lpips==0.1.4
